change id from int to long

+++ src/me/nichel/sportcenter/dao/CustomerDAO.java
@@ -7,11 +7,11 @@ import me.nichel.sportcenter.model.Customer;
 public interface CustomerDAO {
 	public void add(Customer customer);
 
-	public void delete(int id);
+	public void delete(long id);
 
 	public void update(Customer customer);
 
 	public List<Customer> getAll();
 
-	public Customer get(int id);
+	public Customer get(long id);
 }
